bot/requirements.txt
aiogram==2.23.1
redis==4.4.0
requests==2.28.1
PyYAML==5.4.1